--- documents/entity/service.go
@@ -1,226 +1,0 @@
-package entity
-
-/*
-import (
-	"context"
-
-	"github.com/centrifuge/centrifuge-protobufs/gen/go/coredocument"
-	"github.com/centrifuge/go-centrifuge/contextutil"
-	"github.com/centrifuge/go-centrifuge/documents"
-	"github.com/centrifuge/go-centrifuge/errors"
-	cliententitypb "github.com/centrifuge/go-centrifuge/protobufs/gen/go/entity"
-	"github.com/centrifuge/go-centrifuge/queue"
-	"github.com/centrifuge/go-centrifuge/transactions"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-)
-
-// Service defines specific functions for entity
-type Service interface {
-	documents.Service
-
-	// DeriverFromPayload derives Entity from clientPayload
-	DeriveFromCreatePayload(ctx context.Context, payload *cliententitypb.EntityCreatePayload) (documents.Model, error)
-
-	// DeriveFromUpdatePayload derives entity model from update payload
-	DeriveFromUpdatePayload(ctx context.Context, payload *cliententitypb.EntityUpdatePayload) (documents.Model, error)
-
-	// DeriveEntityData returns the entity data as client data
-	DeriveEntityData(inv documents.Model) (*cliententitypb.EntityData, error)
-
-	// DeriveEntityResponse returns the entity model in our standard client format
-	DeriveEntityResponse(inv documents.Model) (*cliententitypb.EntityResponse, error)
-}
-
-// service implements Service and handles all entity related persistence and validations
-// service always returns errors of type `errors.Error` or `errors.TypedError`
-type service struct {
-	documents.Service
-	repo      documents.Repository
-	queueSrv  queue.TaskQueuer
-	txManager transactions.Manager
-}
-
-// DefaultService returns the default implementation of the service.
-func DefaultService(
-	srv documents.Service,
-	repo documents.Repository,
-	queueSrv queue.TaskQueuer,
-	txManager transactions.Manager,
-) Service {
-	return service{
-		repo:      repo,
-		queueSrv:  queueSrv,
-		txManager: txManager,
-		Service:   srv,
-	}
-}
-
-// DeriveFromCoreDocument takes a core document model and returns an entity
-func (s service) DeriveFromCoreDocument(cd coredocumentpb.CoreDocument) (documents.Model, error) {
-	inv := new(Entity)
-	err := inv.UnpackCoreDocument(cd)
-	if err != nil {
-		return nil, errors.NewTypedError(documents.ErrDocumentUnPackingCoreDocument, err)
-	}
-
-	return inv, nil
-}
-
-// UnpackFromCreatePayload initializes the model with parameters provided from the rest-api call
-func (s service) DeriveFromCreatePayload(ctx context.Context, payload *cliententitypb.EntityCreatePayload) (documents.Model, error) {
-	if payload == nil || payload.Data == nil {
-		return nil, documents.ErrDocumentNil
-	}
-
-	did, err := contextutil.AccountDID(ctx)
-	if err != nil {
-		return nil, documents.ErrDocumentConfigAccountID
-	}
-
-	entityModel := new(Entity)
-	err = entityModel.InitEntityInput(payload, did.String())
-	if err != nil {
-		return nil, errors.NewTypedError(documents.ErrDocumentInvalid, err)
-	}
-
-	return entityModel, nil
-}
-
-// validateAndPersist validates the document, calculates the data root, and persists to DB
-func (s service) validateAndPersist(ctx context.Context, old, new documents.Model, validator documents.Validator) (documents.Model, error) {
-	selfDID, err := contextutil.AccountDID(ctx)
-	if err != nil {
-		return nil, errors.NewTypedError(documents.ErrDocumentConfigAccountID, err)
-	}
-
-	inv, ok := new.(*Entity)
-	if !ok {
-		return nil, errors.NewTypedError(documents.ErrDocumentInvalidType, errors.New("unknown document type: %T", new))
-	}
-
-	// validate the entity
-	err = validator.Validate(old, inv)
-	if err != nil {
-		return nil, errors.NewTypedError(documents.ErrDocumentInvalid, err)
-	}
-
-	// we use CurrentVersion as the id since that will be unique across multiple versions of the same document
-	err = s.repo.Create(selfDID[:], inv.CurrentVersion(), inv)
-	if err != nil {
-		return nil, errors.NewTypedError(documents.ErrDocumentPersistence, err)
-	}
-
-	return inv, nil
-}
-
-// Create takes and entity model and does required validation checks, tries to persist to DB
-func (s service) Create(ctx context.Context, inv documents.Model) (documents.Model, transactions.TxID, chan bool, error) {
-	selfDID, err := contextutil.AccountDID(ctx)
-	if err != nil {
-		return nil, transactions.NilTxID(), nil, errors.NewTypedError(documents.ErrDocumentConfigAccountID, err)
-	}
-
-	inv, err = s.validateAndPersist(ctx, nil, inv, CreateValidator())
-	if err != nil {
-		return nil, transactions.NilTxID(), nil, err
-	}
-
-	txID := contextutil.TX(ctx)
-	txID, done, err := documents.CreateAnchorTransaction(s.txManager, s.queueSrv, selfDID, txID, inv.CurrentVersion())
-	if err != nil {
-		return nil, transactions.NilTxID(), nil, err
-	}
-	return inv, txID, done, nil
-}
-
-// Update finds the old document, validates the new version and persists the updated document
-func (s service) Update(ctx context.Context, new documents.Model) (documents.Model, transactions.TxID, chan bool, error) {
-	selfDID, err := contextutil.AccountDID(ctx)
-	if err != nil {
-		return nil, transactions.NilTxID(), nil, errors.NewTypedError(documents.ErrDocumentConfigAccountID, err)
-	}
-
-	old, err := s.GetCurrentVersion(ctx, new.ID())
-	if err != nil {
-		return nil, transactions.NilTxID(), nil, errors.NewTypedError(documents.ErrDocumentNotFound, err)
-	}
-
-	new, err = s.validateAndPersist(ctx, old, new, UpdateValidator())
-	if err != nil {
-		return nil, transactions.NilTxID(), nil, err
-	}
-
-	txID := contextutil.TX(ctx)
-	txID, done, err := documents.CreateAnchorTransaction(s.txManager, s.queueSrv, selfDID, txID, new.CurrentVersion())
-	if err != nil {
-		return nil, transactions.NilTxID(), nil, err
-	}
-	return new, txID, done, nil
-}
-
-// DeriveEntityResponse returns create response from entity model
-func (s service) DeriveEntityResponse(model documents.Model) (*cliententitypb.EntityResponse, error) {
-	data, err := s.DeriveEntityData(model)
-	if err != nil {
-		return nil, err
-	}
-
-	cs, err := model.GetCollaborators()
-	if err != nil {
-		return nil, errors.New("failed to get collaborators: %v", err)
-	}
-
-	var css []string
-	for _, c := range cs {
-		css = append(css, c.String())
-	}
-
-	h := &cliententitypb.ResponseHeader{
-		DocumentId:    hexutil.Encode(model.ID()),
-		VersionId:     hexutil.Encode(model.CurrentVersion()),
-		Collaborators: css,
-	}
-
-	return &cliententitypb.EntityResponse{
-		Header: h,
-		Data:   data,
-	}, nil
-
-}
-
-// DeriveEntityData returns create response from entity model
-func (s service) DeriveEntityData(doc documents.Model) (*cliententitypb.EntityData, error) {
-	inv, ok := doc.(*Entity)
-	if !ok {
-		return nil, documents.ErrDocumentInvalidType
-	}
-
-	return inv.getClientData(), nil
-}
-
-// DeriveFromUpdatePayload returns a new version of the old entity identified by identifier in payload
-func (s service) DeriveFromUpdatePayload(ctx context.Context, payload *cliententitypb.EntityUpdatePayload) (documents.Model, error) {
-	if payload == nil || payload.Data == nil {
-		return nil, documents.ErrDocumentNil
-	}
-
-	// get latest old version of the document
-	id, err := hexutil.Decode(payload.Identifier)
-	if err != nil {
-		return nil, errors.NewTypedError(documents.ErrDocumentIdentifier, errors.New("failed to decode identifier: %v", err))
-	}
-
-	old, err := s.GetCurrentVersion(ctx, id)
-	if err != nil {
-		return nil, err
-	}
-
-	inv := new(Entity)
-	err = inv.PrepareNewVersion(old, payload.Data, payload.Collaborators)
-	if err != nil {
-		return nil, errors.NewTypedError(documents.ErrDocumentPrepareCoreDocument, errors.New("failed to load entity from data: %v", err))
-	}
-
-	return inv, nil
-}
-*/
