--- documents/entity/validator_test.go
@@ -1,53 +1,0 @@
-// +build unit
-
-package entity
-
-/*
-import (
-	"testing"
-
-	"github.com/centrifuge/go-centrifuge/errors"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestFieldValidator_Validate(t *testing.T) {
-	fv := fieldValidator()
-
-	//  nil error
-	err := fv.Validate(nil, nil)
-	assert.Error(t, err)
-	errs := errors.GetErrs(err)
-	assert.Len(t, errs, 1, "errors length must be one")
-	assert.Contains(t, errs[0].Error(), "nil document")
-
-	// unknown type
-	err = fv.Validate(nil, &mockModel{})
-	assert.Error(t, err)
-	errs = errors.GetErrs(err)
-	assert.Len(t, errs, 1, "errors length must be one")
-	assert.Contains(t, errs[0].Error(), "unknown document type")
-
-	// fail
-	err = fv.Validate(nil, new(Entity))
-	assert.Error(t, err)
-	errs = errors.GetErrs(err)
-	assert.Len(t, errs, 1, "errors length must be 2")
-	assert.Contains(t, errs[0].Error(), "currency is invalid")
-
-	// success
-	err = fv.Validate(nil, &Entity{
-		Currency: "EUR",
-	})
-	assert.Nil(t, err)
-}
-
-func TestCreateValidator(t *testing.T) {
-	cv := CreateValidator()
-	assert.Len(t, cv, 1)
-}
-
-func TestUpdateValidator(t *testing.T) {
-	uv := UpdateValidator()
-	assert.Len(t, uv, 2)
-}
-*/
